+++ src/Auth/TokenGeneratorInterface.php
@@ -51,6 +51,8 @@ interface TokenGeneratorInterface
     /**
      * Creates a new authentication token.
      *
+     * @deprecated Use {@see createCustomToken} instead.
+     *
      * @param int|string $id       The user id.
      * @param string     $provider The authentication provider.
      *
@@ -59,4 +61,13 @@ interface TokenGeneratorInterface
      * @return string
      */
     public function createToken($id, $provider);
+
+    /**
+     * Creates a new custom token.
+     *
+     * @param string $uid
+     * @param array $claims
+     * @return mixed
+     */
+    public function createCustomToken($uid, array $claims = []);
 }